Adds a two-database workflow and instructions; gets some migrations working without fully migrating the data
