+++ app/helpers/application_helper.rb
@@ -1,59 +1,57 @@
 module ApplicationHelper
 
-  RAILS_TABLES = %w(
-    schema_migrations
-    assignments
-    defendants
-    organizations
-    defendants_organizations
-    sentences
-    sentence_components
-  )
-
-  def db
-    @db ||= ActiveRecord::Base.connection
+  def oracle_db
+    @oracle_db ||= LegacyRecord.connection
   end
 
-  def all_tables
-    db.tables - RAILS_TABLES
+  def legacy_tables
+    @legacy_tables ||= oracle_db.tables
+  end
+
+  def keep_tables
+    File.read('db/data/tables.keep').each_line.map &:strip
   end
 
   # Gets every tables' column names.
   # Use with #grep to search
   def all_column_names(count=nil)
-    num = count || all_tables.count
-    all_tables.first(num).flat_map.with_index do |table, i|
+    num = count || legacy_tables.count
+    legacy_tables.first(num).flat_map.with_index do |table, i|
       puts "processing table (#{i+1}/#{num}): #{table}"
-      db.columns(table).map(&:name).map do |col|
+      oracle_db.columns(table).map(&:name).map do |col|
         "#{table}.#{col}"
       end
     end
   end
 
   # Get all the models that exist
-  def all_models
-    ApplicationRecord.descendants
+  def legacy_models
+    LegacyRecord.descendants
   end
 
   # Dynamically create a models from a table name, explicitly
   #   setting the table name because the auto-linker doesn't
   #   play well with CRT's naming convention.
   def initialize_model(table_name)
-    Object.const_set(
-      table_name.classify.gsub(/([#|$])/, ''),
-      Class.new(ApplicationRecord) do
-        self.table_name = table_name
-        self.inheritance_column = false # Prevents issues with columns named `type`
-      end
-    )
+    klass_name = table_name.classify.gsub(/([#|$])/, '')
+    begin
+      klass_name.constantize
+    rescue NameError => e
+      Object.const_set(
+        klass_name,
+        Class.new(ApplicationRecord) do
+          self.table_name = table_name
+          self.inheritance_column = false # Prevents issues with columns named `type`
+        end
+      )
+    end
   end
 
   # Dynamically create all the models
   def initialize_models
     erroneous_models = []
-    all_tables.each do |table_name|
+    legacy_tables.each do |table_name|
       begin
-        raise NameError if RAILS_TABLES.include?(table_name)
         initialize_model(table_name)
       rescue NameError => e
         erroneous_models.push([table_name, e])
@@ -63,14 +61,16 @@ module ApplicationHelper
   end
 
   def handle_erroneous_models(model_names)
-    return false unless model_names.any?
+    unless model_names.any?
+      puts "[initialize models] no errors"
+      return true
+    end
     puts "\nThe following tables produced errors, so there are no models for them:\n"
     model_names.each do |table_name, error|
       puts "#{table_name}\t\t#{error.original_message}"
     end
     puts "\n\n"
-
-    true # suppress printing of table names
+    false # suppress printing of table names
   end
 
   def tally_by(attribute_key, first: nil)
@@ -107,8 +107,8 @@ module ApplicationHelper
 
   # Thanks to https://asktom.oracle.com/pls/apex/asktom.search?tag=how-to-calculate-current-db-size
   def db_size
-    physical = db.exec_query("select sum(bytes)/1024/1024 size_in_mb from dba_data_files").rows.first.first
-    allocated = db.exec_query("select sum(bytes)/1024/1024 size_in_mb from dba_segments").rows.first.first
+    physical = oracle_db.exec_query("select sum(bytes)/1024/1024 size_in_mb from dba_data_files").rows.first.first
+    allocated = oracle_db.exec_query("select sum(bytes)/1024/1024 size_in_mb from dba_segments").rows.first.first
     puts <<~MSG
 
       Allocated space: #{(allocated / 1024).round(2)} GB (max: 12 GB)
