+++ app/models/defendant.rb
@@ -1,3 +1,4 @@
-class Defendant < ApplicationRecord
+class Defendant < ModernRecord
+  has_many :sentences
   has_and_belongs_to_many :organizations
 end
