+++ app/models/legacy_record.rb
@@ -1,0 +1,3 @@
+class LegacyRecord < ApplicationRecord
+  self.abstract_class = true
+end
