+++ app/models/modern_record.rb
@@ -1,0 +1,4 @@
+class ModernRecord < ApplicationRecord
+  self.abstract_class = true
+  connects_to database: { reading: :modern, writing: :modern }
+end
