+++ db/migrate/20230727163712_load_defendants.rb
@@ -1,0 +1,11 @@
+class LoadDefendants < ActiveRecord::Migration[7.0]
+  def change
+    reversible do |direction|
+      direction.up do
+        ModernRecord.transaction {
+          DataTransform::LoadDefendants.test
+        }
+      end
+    end
+  end
+end
