+++ db/migrate/20230731184340_create_sentences.rb
@@ -1,33 +1,26 @@
 class CreateSentences < ActiveRecord::Migration[7.0]
   def change
-    ActiveRecord::Base.transaction do
-      create_table :sentences, force: :cascade do |t|
-        t.references :defendant, null: false, foreign_key: true
-        # t.bigint :matter_id, null: false
-        t.datetime :sentencing_date
-        t.timestamps
-      end
+    create_table :sentences, force: :cascade do |t|
+      t.references :defendant, null: false, foreign_key: true
+      # t.bigint :matter_id, null: false
+      t.datetime :sentencing_date
+      t.timestamps
+    end
 
-      # add_foreign_key :sentences, :crdmain, column: :matter_id, primary_key: :matter_no
+    # add_foreign_key :sentences, :crdmain, column: :matter_id, primary_key: :matter_no
 
-      create_table :sentence_components, force: :cascade do |t|
-        t.integer :amount
-        t.text :comments
-        t.references :sentence, null: false, foreign_key: true
-        t.integer :type # enum
-        t.integer :duration_unit # enum
-        t.integer :duration_quantity
-        t.integer :duration
-        t.boolean :life_in_prison, null: true, default: nil
-        t.boolean :time_served, null: true, default: nil
+    create_table :sentence_components, force: :cascade do |t|
+      t.integer :amount
+      t.text :comments
+      t.references :sentence, null: false, foreign_key: true
+      t.integer :type # enum
+      t.integer :duration_unit # enum
+      t.integer :duration_quantity
+      t.integer :duration
+      t.boolean :life_in_prison, null: true, default: nil
+      t.boolean :time_served, null: true, default: nil
 
-        t.timestamps
-      end
-      reversible do |direction|
-        direction.up do
-          DataTransform::LoadSentences.perform
-        end
-      end
+      t.timestamps
     end
   end
 end
