+++ db/migrate/20230731184341_load_sentences.rb
@@ -1,0 +1,11 @@
+class LoadSentences < ActiveRecord::Migration[7.0]
+  def change
+    reversible do |direction|
+      direction.up do
+        ModernRecord.transaction {
+          DataTransform::LoadSentences.test
+        }
+      end
+    end
+  end
+end
